Update style.scss to add padding to code sections
The four values for `padding` are `top - left - bottom - right`, so this change adds a small amount of padding on all four edges, and a bit more on the left. IMHO this slightly improves the appearance of the code samples. I read somewhere that using `em` is the way to allow styles to increase and decrease in size gracefully as the user changes their browser's text size.

(Caveat: I'm really not an expert in all this, so I can't guarantee that I've done the right thing... :) )