Create the SciML Small Grants Program
This is a proposal for the creation of a SciML Small Grants program. It is designed to give an incentive for new contributors to solve major issues that would catalyze further development, or help with user experience stumbling blocks that have been consistently ignored over the last few years. For example, updating benchmarks or projects that would improve run time, the kinds of things that are not research projects but fundamental to the growth of the ecosystem.

The design of the small grants program is different from a bounty program because we want to encourage the development of new contributors to the ecosystem. Thus the setup of "winner takes all" in a bug bounty, i.e. 2 people try and solve the same problem and only the first gets the pay out, it's not what we are looking for. Additionally, we want to make this be administered from the top down, i.e. we want to pick what the issues to solve are because we know what would truly achieve these aims, whereas bug bounties can hit some odd issues that maintainers aren't interested in helping with. But with that we should have some guarantees of reviews. Thus the proposed structure of this system is to consult with the crew to get a yes/no to block out a bit of exclusive time to solve the problems, and if completed, get the bounty.

I think such a program can have a lot of potential and is the kind of thing we can seek further donations to continue, both here and in the JuliaLang organization itself.

But in order to kick this off, I think it needs to at least be given the trial. As such, I plan to donate the funds to SciML to cover these projects myself, and plan to commit to being a reviewer on these projects. If this trial run is successful, I think we can share this success to try and capture more community donations and grow the system some more. But for now it's a bit limited based off of what I could save up, and I'm focusing it on just the absolute core issues that would help startup times the most.

I'm seeking comments, nothing is completely set in stone, this is just what I have been able to cook up in my head over the last few months and hopefully it all makes sense.