Refresh the SciML Small Grants program
This is a grouping of changes to the administration of the system:

1. Moving the administration out of email and to Github. This should make it easier for others to get involved and help out with the administration, and leaves a better history.
2. Changing to only require a majority of the steering council to approve a project. The previous system left this ambiguous so we always waited for all, but that seems unnecessary.
3. Clarity on how others can add projects to the system.
4. Refreshing projects for some of the updates of the ecosystem. Some projects were eliminated, and I'm adding a few projects which can be funded.

Note that if there's other high priority projects, there are some funds from the CZI grant which went unused that we could use for some higher priority projects. We should discuss whether there are any to add here.

Because this is also including substantive changes to the system, it requires an approval by all members of the steering council.