Declare SciMLOperators completed and add more information on payout
The SciMLOperators v1.0 project was completed in https://github.com/SciML/SciMLOperators.jl/pull/269. I did a few things to finish it up but most of the work should be attributed to @divital-coder as I just did a few things in a few hours to just handle the last little bits and get it done, he did all of the big nasty work. Thanks! This is a major step we've been waiting on for awhile.